Rearrange texel tuning related code.
