+++ src/RubiChess.vcxproj
@@ -135,6 +135,7 @@
     <ClCompile Include="nnue.cpp" />
     <ClCompile Include="search.cpp" />
     <ClCompile Include="tbprobe.cpp" />
+    <ClCompile Include="texel.cpp" />
     <ClCompile Include="transposition.cpp" />
     <ClCompile Include="uci.cpp" />
     <ClCompile Include="utils.cpp" />
